# Faster State Iteration