Previous state iterator called `state_from_index` on each iteration which in turn creates `CartesianIndices` for which only a single element is used.

In this PR, the temporary `CartesianIndices` allocation at each iteration is removed by allocating at the very beginning then including it in the iteration state.

The following benchmark is used for comparison:
```julia
rs = RockSamplePOMDP(5,7)
@benchmark collect(states(rs))
```

Previously, we get mean time of: `3.839 ms ±  2.036 ms`
Now, we get: `731.324 μs ± 643.553 μs`

This is a ~5x speedup, and the new array of collected states is equivalent to the vector created with the old iteration method (verified only on `RockSamplePOMDP(5,7)`)