@WhiffleFish 2021-12-19
Benchmark above brought down to `361.415 μs ± 341.099 μs` by making rock state static vector creation nonallocating (with `@view`), and faster "closed-form" solution to `CartesianIndices(v)[idx]`.