@WhiffleFish 2021-12-23
Yeah you're entirely right. My changes were made under the impression that `CartesianIndices` actually allocated a whole vector filled with `CartesianIndex`, but that's clearly wrong. One of the problems was, as you stated, splatting the rocks_dim Vector into the `CartesianIndices` constructor, but the other big issue was the rocks static array:
```julia
rocks = SVector{K, Bool}([(s[i] - 1) for i=3:K+2])
```
The previous code allocated space for a regular base julia vector before converting to static vector. The new code performs this same calculation without the comprehension that was kicking up a lot of garbage collection.